+++ concrete_datastore/concrete/apps.py
@@ -1,14 +1,18 @@
 # coding: utf-8
+import os
 import hashlib
 import logging
+from importlib import import_module
 from collections import defaultdict
 
 from django.utils.translation import ugettext_lazy as _
 from django.apps import AppConfig
+from django.conf import settings
 from django.apps import apps
 
 
 logger_archive_users = logging.getLogger('archive-concrete-users')
+logger_update_pendulum = logging.getLogger('update-pendulum-v2')
 
 
 def email_to_username(email):  # nosec
@@ -52,12 +56,56 @@ def purge_users_with_wrong_usernames(user_model):
                 archive_legacy(user=user, email_list=email_list)
 
 
+def alter_migration_content(file):
+    file_changed = False
+    with open(file, 'rb') as fd:
+        content = fd.read()
+        if b'import pendulum.pendulum' in content:
+            file_changed = True
+            content = content.replace(
+                b'pendulum.pendulum.Pendulum', b'pendulum'
+            ).replace(b'import pendulum.pendulum', b'import pendulum')
+    if file_changed is True:
+        logger_update_pendulum.debug(
+            f'Updating pendulum import for migration file {file}'
+        )
+        with open(file, 'wb') as fd:
+            fd.write(content)
+
+
+def get_migration_files():
+    migrations_dir = set()
+    for migration_module in settings.MIGRATION_MODULES.values():
+        module = import_module(migration_module)
+        migrations_dir.add(os.path.dirname(module.__file__))
+
+    migration_files = []
+    for migration_dir in migrations_dir:
+        for root, dirs, files in os.walk(migration_dir):
+            migration_files.extend(
+                os.path.join(root, f) for f in files if f.endswith('.py')
+            )
+            break
+
+    return migration_files
+
+
+def update_pendulum_for_migrations():
+    migration_files = get_migration_files()
+    for migration_file in migration_files:
+        alter_migration_content(file=migration_file)
+
+
 class Config(AppConfig):
     name = 'concrete_datastore.concrete'
     label = 'concrete'
     verbose_name = _('NS concrete')
 
     def ready(self):
+        logger_update_pendulum.debug(
+            'Checking for pendulum updates in migrations'
+        )
+        update_pendulum_for_migrations()
         user_model = apps.get_model('concrete', 'User')
         try:
             purge_users_with_wrong_usernames(user_model=user_model)