+++ tests/unittest_settings.py
@@ -859,4 +859,8 @@ META_MODEL_DEFINITIONS = (
 
 
 DISABLED_MODELS = ("EntityDividerModel",)
-EMAIL_HOST = ''
+
+EMAIL_HOST = os.environ.get("EMAIL_HOST", 'localhost')
+EMAIL_HOST_USER = os.environ.get("EMAIL_HOST_USER", '')
+EMAIL_HOST_PASSWORD = os.environ.get("EMAIL_HOST_PASSWORD", '')
+EMAIL_PORT = os.environ.get("EMAIL_PORT", 1025)